+++ razorboard/src/app/polled-view.ts
@@ -1,7 +1,6 @@
 import { OnInit, OnDestroy } from '@angular/core';
 import { EMPTY, BehaviorSubject, timer, interval, merge } from 'rxjs';
 import { catchError, startWith, switchMap } from 'rxjs/operators';
-import { ToastrService } from 'ngx-toastr';
 import { Params, Router, ActivatedRoute } from '@angular/router';
 
 import { ApiResponse } from './models/apiresponse.model';
@@ -14,7 +13,6 @@ export abstract class PolledView implements OnInit, OnDestroy {
 
   constructor(
     protected razorApi: RazorapiService,
-    protected toastr: ToastrService,
     protected route: ActivatedRoute,
     protected router: Router,
   ) { }
@@ -30,7 +28,6 @@ export abstract class PolledView implements OnInit, OnDestroy {
       switchMap(() => this.getData().pipe(
         // This inner pipe avoids the observable to be completed when an HTTP error arises
         catchError((err, caugth) => {
-          this.toastr.error(err.message, 'Unable to fetch data');
           return EMPTY;
         }),
       )),