+++ razorboard/src/app/top-bar/top-bar.component.ts
@@ -1,11 +1,11 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { FormControl, FormGroup, Validators } from '@angular/forms';
-import { NgbModal, ModalDismissReasons } from '@ng-bootstrap/ng-bootstrap';
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { Subject } from 'rxjs';
-import { faCogs } from '@fortawesome/free-solid-svg-icons';
+import { faCogs, faPlay, faPause } from '@fortawesome/free-solid-svg-icons';
 
+import { SettingsModalComponent } from '../settings-modal/settings-modal.component';
 import { RazorapiService } from '../razorapi.service';
-import { HttpLoadingService } from '../http-loading.service';
+import { HttpEventsService } from '../http-events.service';
 
 @Component({
   selector: 'app-top-bar',
@@ -15,11 +15,12 @@ import { HttpLoadingService } from '../http-loading.service';
 export class TopBarComponent implements OnInit {
   @Input() title: string;
   activeItem: string;
-  connForm: FormGroup;
 
-  httpLoading: Subject<boolean> = this.loaderService.loading;
+  httpLoading: Subject<boolean> = this.httpEventsService.loading;
 
   faCogs = faCogs;
+  faPlay = faPlay;
+  faPause = faPause;
 
   menuItems = [
     {
@@ -58,41 +59,19 @@ export class TopBarComponent implements OnInit {
 
   constructor(
     public razorapiService: RazorapiService,
-    private loaderService: HttpLoadingService,
+    private httpEventsService: HttpEventsService,
     private modalService: NgbModal,
   ) { }
 
   ngOnInit() {
-    this.connForm = new FormGroup({
-      endpoint: new FormControl(this.razorapiService.getEndpoint(), [
-        Validators.required,
-        Validators.pattern('^(https?:\/\/)?(www\.)?[a-z0-9]+([\-\.]{1}[a-z0-9]+)*\.[a-z]{2,5}(:[0-9]{1,5})?(\/.*)?$'),
-      ]),
-      username: new FormControl(this.razorapiService.getUsername()),
-      password: new FormControl(this.razorapiService.getPassword()),
-      refresh: new FormControl({value: this.razorapiService.getRefresh(), disabled: !this.razorapiService.getRefreshEnabled()}),
-      refreshEnabled: new FormControl(this.razorapiService.getRefreshEnabled()),
-    });
-    this.connForm.get('refreshEnabled').valueChanges.subscribe(v => {
-      if (v) {
-        this.connForm.get('refresh').enable();
-      } else {
-        this.connForm.get('refresh').disable();
-      }
-    });
   }
 
-  open(content) {
-    this.modalService.open(content, {ariaLabelledBy: 'settings-modal'}).result.then((result) => {
-      if (result === 'save') {
-        this.razorapiService.connect(
-          this.connForm.value.endpoint,
-          this.connForm.value.username,
-          this.connForm.value.password,
-          this.connForm.value.refresh,
-          this.connForm.value.refreshEnabled,
-        );
-      }
-    });
+  open() {
+    const modalRef = this.modalService.open(SettingsModalComponent);
+    modalRef.componentInstance.title = this.title;
+  }
+
+  switchRefresh(value: boolean) {
+    this.razorapiService.setRefreshEnabled(!value);
   }
 }